refactor event tools
