+++ packages/ai-constructs/src/conversation/runtime/event-tools-provider/event_tools_provider.test.ts
@@ -1,0 +1,5 @@
+import { describe, it } from 'node:test';
+
+void describe('bedrock converse adapter', () => {
+  void it('should construct appsync query tool from event definition', async () => {});
+});
