+++ packages/ai-constructs/src/conversation/runtime/event-tools-provider/index.ts
@@ -1,0 +1,3 @@
+import { ConversationTurnEventToolsProvider } from './event_tools_provider.js';
+
+export { ConversationTurnEventToolsProvider };
