+++ packages/ai-constructs/src/conversation/runtime/event-tools-provider/types.ts
@@ -1,0 +1,5 @@
+import { ConversationTurnEvent } from '../types';
+
+export type ConversationTurnEventToolConfiguration = NonNullable<
+  ConversationTurnEvent['toolsConfiguration']
+>['tools'][number];
